LayoutTests/imported/w3c/web-platform-tests/css/css-values/media-progress-serialize.tentative-expected.txt


FAIL 'media-progress(width from 0px to 1000px)' as a specified value should serialize as 'media-progress(width from 0px to 1000px)'. assert_not_equals: 'media-progress(width from 0px to 1000px)' should be valid in opacity. got disallowed value ""
FAIL 'scale(media-progress(width from 0px to 1000px))' as a specified value should serialize as 'scale(media-progress(width from 0px to 1000px))'. assert_not_equals: 'scale(media-progress(width from 0px to 1000px))' should be valid in transform. got disallowed value ""
FAIL 'media-progress(width from 0px to 1000px)' as a computed value should serialize as '0.8'. assert_equals: 'media-progress(width from 0px to 1000px)' and '0.8' should serialize the same in computed values. expected "0.8" but got "1"
FAIL 'scale(media-progress(width from 0px to 1000px))' as a computed value should serialize as 'matrix(0.8, 0, 0, 0.8, 0, 0)'. assert_equals: 'scale(media-progress(width from 0px to 1000px))' and 'matrix(0.8, 0, 0, 0.8, 0, 0)' should serialize the same in computed values. expected "matrix(0.8, 0, 0, 0.8, 0, 0)" but got "none"
FAIL 'calc(0.1 * media-progress(height from 0px to 10em))' as a specified value should serialize as 'calc(0.1 * media-progress(height from 0px to 10em))'. assert_not_equals: 'calc(0.1 * media-progress(height from 0px to 10em))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(0.1 * media-progress(height from 0px to 10em)))' as a specified value should serialize as 'scale(calc(0.1 * media-progress(height from 0px to 10em)))'. assert_not_equals: 'scale(calc(0.1 * media-progress(height from 0px to 10em)))' should be valid in transform. got disallowed value ""
FAIL 'calc(0.1 * media-progress(height from 0px to 10em))' as a computed value should serialize as '0.6'. assert_equals: 'calc(0.1 * media-progress(height from 0px to 10em))' and '0.6' should serialize the same in computed values. expected "0.6" but got "1"
FAIL 'scale(calc(0.1 * media-progress(height from 0px to 10em)))' as a computed value should serialize as 'matrix(0.6, 0, 0, 0.6, 0, 0)'. assert_equals: 'scale(calc(0.1 * media-progress(height from 0px to 10em)))' and 'matrix(0.6, 0, 0, 0.6, 0, 0)' should serialize the same in computed values. expected "matrix(0.6, 0, 0, 0.6, 0, 0)" but got "none"
FAIL 'calc(media-progress(width from 0px to 100px) / 20)' as a specified value should serialize as 'calc(media-progress(width from 0px to 100px) / 20)'. assert_not_equals: 'calc(media-progress(width from 0px to 100px) / 20)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(media-progress(width from 0px to 100px) / 20))' as a specified value should serialize as 'scale(calc(media-progress(width from 0px to 100px) / 20))'. assert_not_equals: 'scale(calc(media-progress(width from 0px to 100px) / 20))' should be valid in transform. got disallowed value ""
FAIL 'calc(media-progress(width from 0px to 100px) / 20)' as a computed value should serialize as '0.4'. assert_equals: 'calc(media-progress(width from 0px to 100px) / 20)' and '0.4' should serialize the same in computed values. expected "0.4" but got "1"
FAIL 'scale(calc(media-progress(width from 0px to 100px) / 20))' as a computed value should serialize as 'matrix(0.4, 0, 0, 0.4, 0, 0)'. assert_equals: 'scale(calc(media-progress(width from 0px to 100px) / 20))' and 'matrix(0.4, 0, 0, 0.4, 0, 0)' should serialize the same in computed values. expected "matrix(0.4, 0, 0, 0.4, 0, 0)" but got "none"
FAIL 'calc(1px * media-progress(height from abs(11em - 10rem) to 110px / 10))' as a computed value should serialize as '590px'. assert_equals: 'calc(1px * media-progress(height from abs(11em - 10rem) to 110px / 10))' and '590px' should serialize the same in computed values. expected "590px" but got "784px"
